+++ app/src/main/java/org/compose_projects/socialLocal/CONSTANTS.kt
@@ -14,20 +14,11 @@
  * limitations under the License.
  */
 
-package org.compose_projects.socialLocal.feature.profile.ui
+package org.compose_projects.socialLocal
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun ProfileScreen() {
-    Box(Modifier.fillMaxSize(), contentAlignment = Alignment.BottomCenter) {
-        Text(text = "Hola ")
-    }
+object CONSTANTS {
+    const val titleHome = "Chat Global"
+    const val titleInbox = "Mensajes Privados"
+    const val titleProfile = "Perfil"
+    const val titleAuth = "Crear Cuenta"
 }
-
