--- app/src/main/java/org/compose_projects/socialLocal/ui/CONSTANTS.kt
@@ -1,8 +1,0 @@
-package org.compose_projects.socialLocal.ui
-
-object CONSTANTS {
-    const val titleHome = "Chat Global"
-    const val titleInbox = "Mensajes Privados"
-    const val titleProfile = "Perfil"
-    const val titleAuth = "Crear Cuenta"
-}
